# Limit binding of case fold search