Fix for #595