@aaronjensen 2016-07-11
Below are profiles with 100 buffers open.

## Typometer

Green: This branch
Blue: Melpa current

![typometer editor latency analyzer 2016-07-10 19-41-12](https://cloud.githubusercontent.com/assets/8588/16718590/439d5f6a-46d6-11e6-8b0c-6753f0c3ef40.png)
![typometer editor latency analyzer 2016-07-10 19-41-23](https://cloud.githubusercontent.com/assets/8588/16718592/49e8cb16-46d6-11e6-99f1-81b0981b3c16.png)

There wasn't a huge difference there, but it's a difference. I still don't know exactly what state emacs needs to be in to make it as bad as it is sometimes, but given the profiles, this looks like a clear win.


## Before

```
- command-execute                                                 652  49%
 - call-interactively                                             652  49%
  - funcall-interactively                                         652  49%
   - self-insert-command                                          637  48%
    - sp--post-self-insert-hook-handler                           619  47%
     - condition-case                                             619  47%
      - if                                                        619  47%
       - progn                                                    619  47%
        - let                                                     619  47%
         - cond                                                   593  45%
          - if                                                    593  45%
           - setq                                                 593  45%
            - progn                                               593  45%
             - let                                                415  31%
              - cdr                                               415  31%
               - sp--all-pairs-to-insert                          415  31%
                - let                                             415  31%
                 - let                                            415  31%
                  - if                                            314  23%
                   - let                                          314  23%
                    - let                                         314  23%
                     - let                                        314  23%
                      - while                                     314  23%
                       - let                                      314  23%
                        - if                                      314  23%
                         - funcall                                313  23%
                          - sp--looking-back-p                    283  21%
                           - let                                  283  21%
                            - unwind-protect                      275  20%
                             - progn                              274  20%
                              - sp--looking-back                  274  20%
                               - let                              238  18%
                                - if                              238  18%
                                 - save-excursion                 225  17%
                                  - while                         225  17%
                                   - sp--looking-at                216  16%
                                      let                          20   1%
                                     if                             5   0%
                                     and                            1   0%
                          + sp--strict-regexp-quote                 29   2%
                  + let                                           101   7%
             + sp-insert-pair                                     159  12%
             + sp-skip-closing-pair                                18   1%
         + unwind-protect                                          23   1%
    + flycheck-handle-change                                        7   0%
    + js2-mode-edit                                                 5   0%
    + jit-lock-after-change                                         1   0%
   + evil-normal-state                                              8   0%
   + profiler-report                                                4   0%
   + evil-insert                                                    1   0%
+ redisplay_internal (C function)                                 210  15%
+ timer-event-handler                                             195  14%
+ flyspell-post-command-hook                                      193  14%
+ ...                                                              25   1%
+ url-http-generic-filter                                          17   1%
+ evil-escape-pre-command-hook                                      6   0%
+ winner-save-old-configurations                                    5   0%
+ evil-repeat-post-hook                                             3   0%
+ tern-post-command                                                 3   0%
+ flycheck-handle-signal                                            2   0%
+ sp--save-pre-command-state                                        1   0%
+ evil-repeat-pre-hook                                              1   0%
+ hl-paren-initiate-highlight                                       1   0%
```

### Before Reversed

```
+ sp--looking-at                                                  201  15%
  flyspell-post-command-hook                                      135  10%
+ syntax-ppss                                                     124   9%
+ buffer-stale--default-function                                   89   6%
+ let                                                              50   3%
+ sp--wrap-regexp                                                  42   3%
+ apply                                                            36   2%
+ sp--looking-back                                                 36   2%
+ get-frame-persp                                                  32   2%
+ ispell-decode-string                                             32   2%
+ eyebrowse--get                                                   28   2%
+ goto-address-fontify                                             28   2%
+ auto-revert-buffers                                              25   1%
+ intern                                                           25   1%
+ progn                                                            24   1%
+ if                                                               22   1%
+ setq                                                             20   1%
+ nth                                                              19   1%
```

## After

```
- command-execute                                                 254  26%
 - call-interactively                                             254  26%
  - funcall-interactively                                         254  26%
   - self-insert-command                                          243  25%
    - sp--post-self-insert-hook-handler                           240  24%
     - condition-case                                             240  24%
      - if                                                        240  24%
       - progn                                                    240  24%
        - if                                                      240  24%
         - let                                                    224  23%
          - let                                                   224  23%
           - cond                                                 224  23%
            - if                                                  224  23%
             - setq                                               224  23%
              - progn                                             224  23%
               - sp-insert-pair                                   153  15%
                - if                                              153  15%
                 - let*                                           153  15%
                  - unwind-protect                                150  15%
                   - progn                                        150  15%
                    - sp--pair-to-insert                          150  15%
                     - let                                        150  15%
                      - sp--all-pairs-to-insert                   150  15%
                       - let                                      150  15%
                        - let                                     150  15%
                         - let                                    115  11%
                          - while                                 115  11%
                           - let                                  115  11%
                            - if                                  115  11%
                             - sp--do-action-p                    114  11%
                              - let*                              110  11%
                               - if                                46   4%
                                - sp-point-in-string                 46   4%
                                 - condition-case                  46   4%
                                  - progn                          46   4%
                                   - save-excursion                 37   3%
                                    - nth                          37   3%
                                       syntax-ppss                 36   3%
                               + cond                              45   4%
                               + sp-get-pair                       18   1%
                              + setq                                4   0%
                             + progn                                1   0%
                         + if                                      35   3%
                  + if                                              3   0%
               + let                                               58   6%
               + sp-skip-closing-pair                              13   1%
    + js2-mode-edit                                                 2   0%
      flycheck-handle-change                                        1   0%
   + evil-normal-state                                              6   0%
   + profiler-report                                                3   0%
   + evil-insert                                                    1   0%
   + mouse-drag-region                                              1   0%
+ redisplay_internal (C function)                                 232  24%
+ flyspell-post-command-hook                                      193  19%
+ timer-event-handler                                             193  19%
+ ...                                                              32   3%
+ url-http-generic-filter                                          25   2%
+ winner-save-old-configurations                                   13   1%
+ evil-escape-pre-command-hook                                      7   0%
+ evil-repeat-post-hook                                             5   0%
+ flycheck-display-error-at-point-soon                              4   0%
+ flycheck-handle-signal                                            2   0%
+ tern-post-command                                                 2   0%
+ sp--save-pre-command-state                                        2   0%
+ hl-paren-initiate-highlight                                       1   0%
+ evil-repeat-pre-hook                                              1   0%
```

### After Reversed

```
+ sit-for                                                         407  39%
+ redisplay                                                       195  19%
+ syntax-ppss                                                      67   6%
+ get-frame-persp                                                  39   3%
+ intern                                                           21   2%
+ eyebrowse--get                                                   17   1%
+ goto-address-fontify                                             16   1%
+ spaceline-ml-main                                                15   1%
+ progn                                                            14   1%
+ if                                                               12   1%
+ setq                                                             12   1%
+ pl/create-or-get-cache                                           11   1%
+ let                                                              10   0%
+ sp--wrap-regexp                                                  10   0%
```